Close authorization holes in web API routes and Lambdas
The Next.js API routes accepted a client-supplied userid and attached
the internal secret, letting any anonymous visitor read or modify any
user's cards, devices, and Spotify player. Three Lambdas were also
reachable without the internal secret.

- Require a NextAuth session in all six proxy routes and derive userid
  from it instead of the request; NextAuth options move to lib/auth.ts
  so routes can call getServerSession
- Restore the commented-out x-internal check in tapCard and getPlayers;
  add the missing one to setPlayer
- registerUser: upsert tokens on re-login (a revoked refresh token
  previously bricked the user permanently) and stop logging the full
  event, which wrote Spotify tokens to CloudWatch
- Stop forwarding Spotify tokens into the browser session; the client
  only uses the user id
