+++ apps/web/app/api/add-device/route.ts
@@ -1,7 +1,15 @@
 import { NextRequest, NextResponse } from 'next/server';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
 
 export async function POST(req: NextRequest) {
-  const { userid, deviceid } = await req.json();
+  const session = await getServerSession(authOptions);
+  const userid = session?.user?.id;
+  if (!userid) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
+  const { deviceid } = await req.json();
   const LAMBDA_URL = process.env.LAMBDA_URL || '';
   const INTERNAL_SECRET = process.env.LAMBDA_SECRET || '';
