+++ apps/web/app/api/get-cards/route.ts
@@ -1,13 +1,15 @@
-import { NextRequest, NextResponse } from 'next/server';
+import { NextResponse } from 'next/server';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
 
 // This API route proxies to your Lambda endpoint
-export async function GET(req: NextRequest) {
-  const userid = req.nextUrl.searchParams.get('userid');
+export async function GET() {
+  const session = await getServerSession(authOptions);
+  const userid = session?.user?.id;
   if (!userid) {
-    return NextResponse.json({ error: 'Missing userid' }, { status: 400 });
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
 
-  // Replace with your deployed Lambda/API Gateway URL
   const LAMBDA_URL = process.env.LAMBDA_URL || '';
   const INTERNAL_SECRET = process.env.LAMBDA_SECRET || '';
