+++ apps/web/app/api/set-player/route.ts
@@ -1,9 +1,17 @@
 import { NextRequest, NextResponse } from 'next/server';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
 
 export async function POST(req: NextRequest) {
+  const session = await getServerSession(authOptions);
+  const userid = session?.user?.id;
+  if (!userid) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   const body = await req.json();
-  const { userid, spotifyPlayerId, spotifyPlayerName } = body;
-  if (!userid || !spotifyPlayerId || !spotifyPlayerName) {
+  const { spotifyPlayerId, spotifyPlayerName } = body;
+  if (!spotifyPlayerId || !spotifyPlayerName) {
     return NextResponse.json({ error: 'Missing fields' }, { status: 400 });
   }
